add client in url

+++ listener/src/main/java/com/workbei/controller/autocreate/AutoCreateV2Controller.java
@@ -16,10 +16,13 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.Validator;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 import java.util.Map;
 
-import static com.workbei.util.LogFormatter.*;
+import static com.workbei.util.LogFormatter.LogEvent;
+import static com.workbei.util.LogFormatter.getKV;
 
 /**
  * @author Wallace Mao
@@ -31,11 +34,11 @@ import static com.workbei.util.LogFormatter.*;
         position = 0,
         produces = "application/json",
         consumes = "application/json",
-        tags = "tokenAuth")
-@RestController("autoCreateController")
+        tags = "tokenAuthV2")
+@RestController("autoCreateV2Controller")
 @RequestMapping("/v3w/tokenAuth")
-public class AutoCreateController {
-    private static final Logger bizLogger = LoggerFactory.getLogger(AutoCreateController.class);
+public class AutoCreateV2Controller {
+    private static final Logger bizLogger = LoggerFactory.getLogger(AutoCreateV2Controller.class);
 
     private ValidateChecker checker;
     private AutoCreateService autoCreateService;
@@ -47,7 +50,7 @@ public class AutoCreateController {
     private Validator updateUserValidator;
 
     @Autowired
-    public AutoCreateController(ValidateChecker checker, AutoCreateService autoCreateService) {
+    public AutoCreateV2Controller(ValidateChecker checker, AutoCreateService autoCreateService) {
         this.checker = checker;
         this.autoCreateService = autoCreateService;
     }
@@ -94,9 +97,12 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PostMapping("/team")
+    @PostMapping("/client/{client}/team")
     public AutoCreateTeamVO createTeam(
+            @PathVariable("client") String client,
             @RequestBody AutoCreateTeamVO autoCreateTeamVO,
+            HttpServletRequest request,
+            HttpServletResponse response,
             BindingResult result
     ) {
         bizLogger.info(LogFormatter.format(
@@ -104,6 +110,12 @@ public class AutoCreateController {
                 "createTeam",
                 getKV("autoCreateTeamVO", autoCreateTeamVO)
         ));
+        if (autoCreateTeamVO.getClient() == null) {
+            autoCreateTeamVO.setClient(makeClient(client));
+        }
+        if (!checkClient(request, response, autoCreateTeamVO.getClient())) {
+            return null;
+        }
         checker.check(autoCreateTeamVO, createTeamValidator, result);
         autoCreateService.createTeam(autoCreateTeamVO);
         return autoCreateTeamVO;
@@ -132,9 +144,12 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PostMapping("/department")
+    @PostMapping("/client/{client}/department")
     public AutoCreateDepartmentVO createDepartment(
+            @PathVariable("client") String client,
             @RequestBody AutoCreateDepartmentVO autoCreateDepartmentVO,
+            HttpServletRequest request,
+            HttpServletResponse response,
             BindingResult result
     ) {
         bizLogger.info(LogFormatter.format(
@@ -142,6 +157,12 @@ public class AutoCreateController {
                 "createDepartment",
                 getKV("autoCreateDepartmentVO", autoCreateDepartmentVO)
         ));
+        if (autoCreateDepartmentVO.getClient() == null) {
+            autoCreateDepartmentVO.setClient(makeClient(client));
+        }
+        if (!checkClient(request, response, autoCreateDepartmentVO.getClient())) {
+            return null;
+        }
         checker.check(autoCreateDepartmentVO, createDepartmentValidator, result);
         autoCreateService.createDepartment(autoCreateDepartmentVO);
         return autoCreateDepartmentVO;
@@ -170,10 +191,13 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PutMapping("/department/{outerId}")
+    @PutMapping("/client/{client}/department/{outerId}")
     public Map updateDepartment(
+            @PathVariable("client") String client,
             @PathVariable("outerId") String outerId,
             @RequestBody AutoCreateDepartmentVO autoCreateDepartmentVO,
+            HttpServletRequest request,
+            HttpServletResponse response,
             BindingResult result
     ) {
         bizLogger.info(LogFormatter.format(
@@ -182,6 +206,12 @@ public class AutoCreateController {
                 getKV("outerId", outerId),
                 getKV("autoCreateDepartmentVO", autoCreateDepartmentVO)
         ));
+        if (autoCreateDepartmentVO.getClient() == null) {
+            autoCreateDepartmentVO.setClient(makeClient(client));
+        }
+        if (!checkClient(request, response, autoCreateDepartmentVO.getClient())) {
+            return null;
+        }
         if (autoCreateDepartmentVO.getOuterCombineId() == null) {
             autoCreateDepartmentVO.setOuterCombineId(outerId);
         }
@@ -205,9 +235,12 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @DeleteMapping("/department/{outerId}")
+    @DeleteMapping("/client/{client}/department/{outerId}")
     public Map deleteDepartment(
-            @PathVariable("outerId") String outerId
+            @PathVariable("client") String client,
+            @PathVariable("outerId") String outerId,
+            HttpServletRequest request,
+            HttpServletResponse response
     ) {
         bizLogger.info(LogFormatter.format(
                 LogEvent.START,
@@ -215,8 +248,11 @@ public class AutoCreateController {
                 getKV("outerId", outerId)
         ));
         AutoCreateDepartmentVO autoCreateDepartmentVO = new AutoCreateDepartmentVO();
+        autoCreateDepartmentVO.setClient(makeClient(client));
         autoCreateDepartmentVO.setOuterCombineId(outerId);
-        autoCreateDepartmentVO.setClient(WbConstant.APP_DEFAULT_CLIENT);
+        if (!checkClient(request, response, autoCreateDepartmentVO.getClient())) {
+            return null;
+        }
         autoCreateService.deleteDepartment(autoCreateDepartmentVO);
         return ResponseResult.success();
     }
@@ -244,9 +280,12 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PostMapping("/user")
+    @PostMapping("/client/{client}/user")
     public AutoCreateUserVO createUser(
+            @PathVariable("client") String client,
             @RequestBody AutoCreateUserVO userVO,
+            HttpServletRequest request,
+            HttpServletResponse response,
             BindingResult result
     ) {
         bizLogger.info(LogFormatter.format(
@@ -254,6 +293,12 @@ public class AutoCreateController {
                 "createUser",
                 getKV("userVO", userVO)
         ));
+        if (userVO.getClient() == null) {
+            userVO.setClient(makeClient(client));
+        }
+        if (!checkClient(request, response, userVO.getClient())) {
+            return null;
+        }
         checker.check(userVO, createUserValidator, result);
         autoCreateService.createUser(userVO);
         return userVO;
@@ -282,10 +327,13 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PutMapping("/user/{outerId}")
+    @PutMapping("/client/{client}/user/{outerId}")
     public AutoCreateUserVO updateUser(
+            @PathVariable("client") String client,
             @PathVariable("outerId") String outerId,
             @RequestBody AutoCreateUserVO userVO,
+            HttpServletRequest request,
+            HttpServletResponse response,
             BindingResult result
     ) {
         bizLogger.info(LogFormatter.format(
@@ -294,6 +342,12 @@ public class AutoCreateController {
                 getKV("outerId", outerId),
                 getKV("userVO", userVO)
         ));
+        if (userVO.getClient() == null) {
+            userVO.setClient(makeClient(client));
+        }
+        if (!checkClient(request, response, userVO.getClient())) {
+            return null;
+        }
         if (userVO.getOuterCombineId() == null) {
             userVO.setOuterCombineId(outerId);
         }
@@ -317,10 +371,13 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PutMapping("/user/{outerId}/admin/{admin}")
+    @PutMapping("/client/{client}/user/{outerId}/admin/{admin}")
     public Map updateUserSetAdmin(
+            @PathVariable("client") String client,
             @PathVariable("outerId") String outerId,
-            @PathVariable("admin") Boolean isAdmin
+            @PathVariable("admin") Boolean isAdmin,
+            HttpServletRequest request,
+            HttpServletResponse response
     ) {
         bizLogger.info(LogFormatter.format(
                 LogEvent.START,
@@ -329,10 +386,49 @@ public class AutoCreateController {
                 getKV("admin", isAdmin)
         ));
         AutoCreateUserVO userVO = new AutoCreateUserVO();
+        userVO.setClient(makeClient(client));
         userVO.setOuterCombineId(outerId);
         userVO.setAdmin(isAdmin);
-        userVO.setClient(WbConstant.APP_DEFAULT_CLIENT);
+        if (!checkClient(request, response, userVO.getClient())) {
+            return null;
+        }
         autoCreateService.updateUserSetAdmin(userVO);
+        return ResponseResult.success();
+    }
+
+    @ApiOperation(value = "更新团队的所有管理员", httpMethod = "PUT")
+    @ApiImplicitParams({
+            @ApiImplicitParam(required = true, paramType = "header", name = "Authorization", dataType = "string",
+                    value = "授权token", defaultValue = "abcd"),
+            @ApiImplicitParam(required = true, paramType = "path", name = "outerId", dataType = "string",
+                    value = "团队的outerId"),
+            @ApiImplicitParam(required = true, paramType = "body", name = "adminList", dataType = "list",
+                    value = "管理员列表")
+    })
+    @ApiResponses({
+            @ApiResponse(code = 400, message = "参数错误，未提供正确的参数"),
+            @ApiResponse(code = 401, message = "无权限访问，比如token未提供"),
+            @ApiResponse(code = 500, message = "服务器错误"),
+            @ApiResponse(code = 200, message = "成功"),
+    })
+    @PutMapping("/client/{client}/team/{outerId}/admin")
+    public Map updateTeamAllAdmin(
+            @PathVariable("client") String client,
+            @PathVariable("outerId") String outerId,
+            @RequestBody List<String> adminList,
+            HttpServletRequest request,
+            HttpServletResponse response
+    ) {
+        bizLogger.info(LogFormatter.format(
+                LogEvent.START,
+                "updateTeamAllAdmin",
+                getKV("outerId", outerId),
+                getKV("adminList", adminList)
+        ));
+        if (!checkClient(request, response, client)) {
+            return null;
+        }
+        autoCreateService.updateBatchUserSetAdmin(client, outerId, adminList);
         return ResponseResult.success();
     }
 
@@ -349,9 +445,12 @@ public class AutoCreateController {
             @ApiResponse(code = 500, message = "服务器错误"),
             @ApiResponse(code = 200, message = "成功"),
     })
-    @PutMapping("/user/{outerId}/team/null")
+    @PutMapping("/client/{client}/user/{outerId}/team/null")
     public Map updateUserLeaveTeam(
-            @PathVariable("outerId") String outerId
+            @PathVariable("client") String client,
+            @PathVariable("outerId") String outerId,
+            HttpServletRequest request,
+            HttpServletResponse response
     ) {
         bizLogger.info(LogFormatter.format(
                 LogEvent.START,
@@ -359,9 +458,28 @@ public class AutoCreateController {
                 getKV("outerId", outerId)
         ));
         AutoCreateUserVO userVO = new AutoCreateUserVO();
+        userVO.setClient(makeClient(client));
         userVO.setOuterCombineId(outerId);
-        userVO.setClient(WbConstant.APP_DEFAULT_CLIENT);
+        if (!checkClient(request, response, userVO.getClient())) {
+            return null;
+        }
         autoCreateService.updateUserLeaveTeam(userVO);
         return ResponseResult.success();
+    }
+
+    private String makeClient(String client) {
+        return client == null ? WbConstant.APP_DEFAULT_CLIENT : client;
+    }
+
+    private Boolean checkClient(HttpServletRequest request, HttpServletResponse response, String client) {
+        String outerClient = (String)request.getAttribute("outerClient");
+        if (client == null) {
+            return false;
+        }
+        boolean isValid = client.equals(outerClient);
+        if (!isValid) {
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        }
+        return isValid;
     }
 }
